functions/requirements.txt
azure-functions>=1.11.0
azure-identity>=1.13.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.16.0
azure-cosmos>=4.3.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn==1.5.1
joblib>=1.2.0